Create vision_commons library and dummy vision_tasks to get started.
