+++ vision_layer/vision_tasks/cfg/range.cfg
@@ -14,4 +14,8 @@ gen.add("high_s", int_t, 0, "Higher Bound Of S", 10, 0, 255);
 gen.add("low_v", int_t, 0, "Lower Bound Of V", 10, 0, 255);
 gen.add("high_v", int_t, 0, "Higher Bound Of V", 10, 0, 255);
 
-exit(gen.generate(PACKAGE, "vision_commons", "range"))
+gen.add("opening_closing_mat_point", int_t, 0, "", 1, 1, 7);
+gen.add("opening_iter", int_t, 0, "Opening iterations", 1, 1, 10);
+gen.add("closing_iter", int_t, 0, "Closing iterations", 1, 1, 10);
+
+exit(gen.generate(PACKAGE, "vision_commons", "range"));
